+++ x/evm/statedb/statedb_test.go
@@ -1,23 +1,44 @@
 package statedb_test
 
 import (
+	"errors"
 	"math/big"
 	"testing"
 
+	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
+	"github.com/tendermint/tendermint/libs/log"
+	dbm "github.com/tendermint/tm-db"
+
+	"github.com/cosmos/cosmos-sdk/store/rootmulti"
+	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	authkeeper "github.com/cosmos/cosmos-sdk/x/auth/keeper"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
+	capabilitytypes "github.com/cosmos/cosmos-sdk/x/capability/types"
+	paramstypes "github.com/cosmos/cosmos-sdk/x/params/types"
 	"github.com/ethereum/go-ethereum/common"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/core/vm"
 	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/evmos/ethermint/x/evm/statedb"
+	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
+	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
+
+	evmenc "github.com/evmos/ethermint/encoding"
+
+	"github.com/evmos/ethermint/app"
+	ethermint "github.com/evmos/ethermint/types"
+	evmkeeper "github.com/evmos/ethermint/x/evm/keeper"
+	"github.com/evmos/ethermint/x/evm/statedb"
+	evmtypes "github.com/evmos/ethermint/x/evm/types"
 )
 
 var (
 	address       common.Address   = common.BigToAddress(big.NewInt(101))
 	address2      common.Address   = common.BigToAddress(big.NewInt(102))
 	address3      common.Address   = common.BigToAddress(big.NewInt(103))
-	blockHash     common.Hash      = common.BigToHash(big.NewInt(9999))
+	blockHash     common.Hash      = common.BigToHash(big.NewInt(0))
 	emptyTxConfig statedb.TxConfig = statedb.NewEmptyTxConfig(blockHash)
 )
 
@@ -30,11 +51,13 @@ func (suite *StateDBTestSuite) TestAccount() {
 	value1 := common.BigToHash(big.NewInt(2))
 	key2 := common.BigToHash(big.NewInt(3))
 	value2 := common.BigToHash(big.NewInt(4))
+	txConfig := emptyTxConfig
+	txConfig.TxHash = common.BigToHash(big.NewInt(100))
 	testCases := []struct {
 		name     string
-		malleate func(*statedb.StateDB)
+		malleate func(*statedb.StateDB, sdk.MultiStore)
 	}{
-		{"non-exist account", func(db *statedb.StateDB) {
+		{"non-exist account", func(db *statedb.StateDB, cms sdk.MultiStore) {
 			suite.Require().Equal(false, db.Exist(address))
 			suite.Require().Equal(true, db.Empty(address))
 			suite.Require().Equal(big.NewInt(0), db.GetBalance(address))
@@ -42,17 +65,16 @@ func (suite *StateDBTestSuite) TestAccount() {
 			suite.Require().Equal(common.Hash{}, db.GetCodeHash(address))
 			suite.Require().Equal(uint64(0), db.GetNonce(address))
 		}},
-		{"empty account", func(db *statedb.StateDB) {
+		{"empty account", func(db *statedb.StateDB, cms sdk.MultiStore) {
 			db.CreateAccount(address)
 			suite.Require().NoError(db.Commit())
 
-			keeper := db.Keeper().(*MockKeeper)
-			acct := keeper.accounts[address]
-			suite.Require().Equal(statedb.NewEmptyAccount(), &acct.account)
-			suite.Require().Empty(acct.states)
-			suite.Require().False(acct.account.IsContract())
+			ctx, keeper := newTestKeeper(suite.T(), cms)
+			acct := keeper.GetAccount(ctx, address)
+			suite.Require().Equal(statedb.NewEmptyAccount(), acct)
+			suite.Require().False(acct.IsContract())
 
-			db = statedb.New(sdk.Context{}, keeper, emptyTxConfig)
+			db = statedb.New(ctx, keeper, txConfig)
 			suite.Require().Equal(true, db.Exist(address))
 			suite.Require().Equal(true, db.Empty(address))
 			suite.Require().Equal(big.NewInt(0), db.GetBalance(address))
@@ -60,7 +82,7 @@ func (suite *StateDBTestSuite) TestAccount() {
 			suite.Require().Equal(common.BytesToHash(emptyCodeHash), db.GetCodeHash(address))
 			suite.Require().Equal(uint64(0), db.GetNonce(address))
 		}},
-		{"suicide", func(db *statedb.StateDB) {
+		{"suicide", func(db *statedb.StateDB, cms sdk.MultiStore) {
 			// non-exist account.
 			suite.Require().False(db.Suicide(address))
 			suite.Require().False(db.HasSuicided(address))
@@ -71,10 +93,15 @@ func (suite *StateDBTestSuite) TestAccount() {
 			db.AddBalance(address, big.NewInt(100))
 			db.SetState(address, key1, value1)
 			db.SetState(address, key2, value2)
+			codeHash := db.GetCodeHash(address)
 			suite.Require().NoError(db.Commit())
 
+			ctx, keeper := newTestKeeper(suite.T(), cms)
+
+			suite.Require().NotEmpty(keeper.GetCode(ctx, codeHash))
+
 			// suicide
-			db = statedb.New(sdk.Context{}, db.Keeper(), emptyTxConfig)
+			db = statedb.New(ctx, keeper, txConfig)
 			suite.Require().False(db.HasSuicided(address))
 			suite.Require().True(db.Suicide(address))
 
@@ -88,28 +115,30 @@ func (suite *StateDBTestSuite) TestAccount() {
 
 			suite.Require().NoError(db.Commit())
 
+			ctx, keeper = newTestKeeper(suite.T(), cms)
+
 			// not accessible from StateDB anymore
-			db = statedb.New(sdk.Context{}, db.Keeper(), emptyTxConfig)
+			db = statedb.New(ctx, keeper, txConfig)
 			suite.Require().False(db.Exist(address))
 
 			// and cleared in keeper too
-			keeper := db.Keeper().(*MockKeeper)
-			suite.Require().Empty(keeper.accounts)
-			suite.Require().Empty(keeper.codes)
+			suite.Require().Nil(keeper.GetAccount(ctx, address))
+			// code is not deleted when contract suicided.
+			suite.Require().NotEmpty(keeper.GetCode(ctx, codeHash))
 		}},
 	}
 	for _, tc := range testCases {
 		suite.Run(tc.name, func() {
-			keeper := NewMockKeeper()
-			db := statedb.New(sdk.Context{}, keeper, emptyTxConfig)
-			tc.malleate(db)
+			raw, ctx, keeper := setupTestEnv(suite.T())
+			db := statedb.New(ctx, keeper, txConfig)
+			tc.malleate(db, raw)
 		})
 	}
 }
 
 func (suite *StateDBTestSuite) TestAccountOverride() {
-	keeper := NewMockKeeper()
-	db := statedb.New(sdk.Context{}, keeper, emptyTxConfig)
+	_, ctx, keeper := setupTestEnv(suite.T())
+	db := statedb.New(ctx, keeper, emptyTxConfig)
 	// test balance carry over when overwritten
 	amount := big.NewInt(1)
 
@@ -131,16 +160,13 @@ func (suite *StateDBTestSuite) TestDBError() {
 		name     string
 		malleate func(vm.StateDB)
 	}{
-		{"set account", func(db vm.StateDB) {
-			db.SetNonce(errAddress, 1)
-		}},
-		{"delete account", func(db vm.StateDB) {
-			db.SetNonce(errAddress, 1)
-			suite.Require().True(db.Suicide(errAddress))
+		{"negative balance", func(db vm.StateDB) {
+			db.SubBalance(address, big.NewInt(10))
 		}},
 	}
 	for _, tc := range testCases {
-		db := statedb.New(sdk.Context{}, NewMockKeeper(), emptyTxConfig)
+		_, ctx, keeper := setupTestEnv(suite.T())
+		db := statedb.New(ctx, keeper, emptyTxConfig)
 		tc.malleate(db)
 		suite.Require().Error(db.Commit())
 	}
@@ -150,37 +176,37 @@ func (suite *StateDBTestSuite) TestBalance() {
 	// NOTE: no need to test overflow/underflow, that is guaranteed by evm implementation.
 	testCases := []struct {
 		name       string
-		malleate   func(*statedb.StateDB)
+		malleate   func(*statedb.StateDB, sdk.MultiStore)
 		expBalance *big.Int
 	}{
-		{"add balance", func(db *statedb.StateDB) {
+		{"add balance", func(db *statedb.StateDB, cms sdk.MultiStore) {
 			db.AddBalance(address, big.NewInt(10))
 		}, big.NewInt(10)},
-		{"sub balance", func(db *statedb.StateDB) {
+		{"sub balance", func(db *statedb.StateDB, cms sdk.MultiStore) {
 			db.AddBalance(address, big.NewInt(10))
 			// get dirty balance
 			suite.Require().Equal(big.NewInt(10), db.GetBalance(address))
 			db.SubBalance(address, big.NewInt(2))
 		}, big.NewInt(8)},
-		{"add zero balance", func(db *statedb.StateDB) {
+		{"add zero balance", func(db *statedb.StateDB, cms sdk.MultiStore) {
 			db.AddBalance(address, big.NewInt(0))
 		}, big.NewInt(0)},
-		{"sub zero balance", func(db *statedb.StateDB) {
+		{"sub zero balance", func(db *statedb.StateDB, cms sdk.MultiStore) {
 			db.SubBalance(address, big.NewInt(0))
 		}, big.NewInt(0)},
 	}
 
 	for _, tc := range testCases {
 		suite.Run(tc.name, func() {
-			keeper := NewMockKeeper()
-			db := statedb.New(sdk.Context{}, keeper, emptyTxConfig)
-			tc.malleate(db)
+			raw, ctx, keeper := setupTestEnv(suite.T())
+			db := statedb.New(ctx, keeper, emptyTxConfig)
+			tc.malleate(db, raw)
 
 			// check dirty state
 			suite.Require().Equal(tc.expBalance, db.GetBalance(address))
 			suite.Require().NoError(db.Commit())
 			// check committed balance too
-			suite.Require().Equal(tc.expBalance, keeper.accounts[address].account.Balance)
+			suite.Require().Equal(tc.expBalance, keeper.GetBalance(ctx, address))
 		})
 	}
 }
@@ -224,16 +250,19 @@ func (suite *StateDBTestSuite) TestState() {
 
 	for _, tc := range testCases {
 		suite.Run(tc.name, func() {
-			keeper := NewMockKeeper()
-			db := statedb.New(sdk.Context{}, keeper, emptyTxConfig)
+			raw, ctx, keeper := setupTestEnv(suite.T())
+			db := statedb.New(ctx, keeper, emptyTxConfig)
 			tc.malleate(db)
 			suite.Require().NoError(db.Commit())
 
 			// check committed states in keeper
-			suite.Require().Equal(tc.expStates, keeper.accounts[address].states)
+			ctx, keeper = newTestKeeper(suite.T(), raw)
+			for k, v := range tc.expStates {
+				suite.Require().Equal(v, keeper.GetState(ctx, address, k))
+			}
 
 			// check ForEachStorage
-			db = statedb.New(sdk.Context{}, keeper, emptyTxConfig)
+			db = statedb.New(ctx, keeper, emptyTxConfig)
 			collected := CollectContractStorage(db)
 			if len(tc.expStates) > 0 {
 				suite.Require().Equal(tc.expStates, collected)
@@ -265,8 +294,8 @@ func (suite *StateDBTestSuite) TestCode() {
 
 	for _, tc := range testCases {
 		suite.Run(tc.name, func() {
-			keeper := NewMockKeeper()
-			db := statedb.New(sdk.Context{}, keeper, emptyTxConfig)
+			raw, ctx, keeper := setupTestEnv(suite.T())
+			db := statedb.New(ctx, keeper, emptyTxConfig)
 			tc.malleate(db)
 
 			// check dirty state
@@ -277,7 +306,12 @@ func (suite *StateDBTestSuite) TestCode() {
 			suite.Require().NoError(db.Commit())
 
 			// check again
-			db = statedb.New(sdk.Context{}, keeper, emptyTxConfig)
+			//db = statedb.New(sdk.Context{}, keeper, emptyTxConfig)
+			//suite.Require().Equal(tc.expCode, db.GetCode(address))
+			//suite.Require().Equal(len(tc.expCode), db.GetCodeSize(address))
+			//suite.Require().Equal(tc.expCodeHash, db.GetCodeHash(address))
+			ctx, keeper = newTestKeeper(suite.T(), raw)
+			db = statedb.New(ctx, keeper, emptyTxConfig)
 			suite.Require().Equal(tc.expCode, db.GetCode(address))
 			suite.Require().Equal(len(tc.expCode), db.GetCodeSize(address))
 			suite.Require().Equal(tc.expCodeHash, db.GetCodeHash(address))
@@ -330,8 +364,8 @@ func (suite *StateDBTestSuite) TestRevertSnapshot() {
 	}
 	for _, tc := range testCases {
 		suite.Run(tc.name, func() {
-			ctx := sdk.Context{}
-			keeper := NewMockKeeper()
+			raw, ctx, keeper := setupTestEnv(suite.T())
+			ctx = ctx.WithEventManager(sdk.NewEventManager())
 
 			{
 				// do some arbitrary changes to the storage
@@ -344,7 +378,8 @@ func (suite *StateDBTestSuite) TestRevertSnapshot() {
 				suite.Require().NoError(db.Commit())
 			}
 
-			originalKeeper := keeper.Clone()
+			originalState := cloneRawState(suite.T(), raw)
+			ctx, keeper = newTestKeeper(suite.T(), raw)
 
 			// run test
 			db := statedb.New(ctx, keeper, emptyTxConfig)
@@ -352,14 +387,15 @@ func (suite *StateDBTestSuite) TestRevertSnapshot() {
 			tc.malleate(db)
 			db.RevertToSnapshot(rev)
 
-			// check empty states after revert
+			// check empty states after Revert
 			suite.Require().Zero(db.GetRefund())
 			suite.Require().Empty(db.Logs())
 
 			suite.Require().NoError(db.Commit())
 
-			// check keeper should stay the same
-			suite.Require().Equal(originalKeeper, keeper)
+			newState := cloneRawState(suite.T(), raw)
+			// check the committed state stays the same
+			suite.Require().Equal(originalState, newState)
 		})
 	}
 }
@@ -369,7 +405,8 @@ func (suite *StateDBTestSuite) TestNestedSnapshot() {
 	value1 := common.BigToHash(big.NewInt(1))
 	value2 := common.BigToHash(big.NewInt(2))
 
-	db := statedb.New(sdk.Context{}, NewMockKeeper(), emptyTxConfig)
+	_, ctx, keeper := setupTestEnv(suite.T())
+	db := statedb.New(ctx, keeper, emptyTxConfig)
 
 	rev1 := db.Snapshot()
 	db.SetState(address, key, value1)
@@ -386,12 +423,14 @@ func (suite *StateDBTestSuite) TestNestedSnapshot() {
 }
 
 func (suite *StateDBTestSuite) TestInvalidSnapshotId() {
-	db := statedb.New(sdk.Context{}, NewMockKeeper(), emptyTxConfig)
+	_, ctx, keeper := setupTestEnv(suite.T())
+	db := statedb.New(ctx, keeper, emptyTxConfig)
 	suite.Require().Panics(func() {
 		db.RevertToSnapshot(1)
 	})
 }
 
+// TODO: need to fix suite.go:87: test panicked: the EVM module account has not been set
 func (suite *StateDBTestSuite) TestAccessList() {
 	value1 := common.BigToHash(big.NewInt(1))
 	value2 := common.BigToHash(big.NewInt(2))
@@ -458,7 +497,8 @@ func (suite *StateDBTestSuite) TestAccessList() {
 	}
 
 	for _, tc := range testCases {
-		db := statedb.New(sdk.Context{}, NewMockKeeper(), emptyTxConfig)
+		_, ctx, keeper := setupTestEnv(suite.T())
+		db := statedb.New(ctx, keeper, emptyTxConfig)
 		tc.malleate(db)
 	}
 }
@@ -471,7 +511,8 @@ func (suite *StateDBTestSuite) TestLog() {
 		txHash,
 		1, 1,
 	)
-	db := statedb.New(sdk.Context{}, NewMockKeeper(), txConfig)
+	_, ctx, keeper := setupTestEnv(suite.T())
+	db := statedb.New(ctx, keeper, txConfig)
 	data := []byte("hello world")
 	db.AddLog(&ethtypes.Log{
 		Address:     address,
@@ -523,7 +564,8 @@ func (suite *StateDBTestSuite) TestRefund() {
 		}, 0, true},
 	}
 	for _, tc := range testCases {
-		db := statedb.New(sdk.Context{}, NewMockKeeper(), emptyTxConfig)
+		_, ctx, keeper := setupTestEnv(suite.T())
+		db := statedb.New(ctx, keeper, emptyTxConfig)
 		if !tc.expPanic {
 			tc.malleate(db)
 			suite.Require().Equal(tc.expRefund, db.GetRefund())
@@ -541,8 +583,8 @@ func (suite *StateDBTestSuite) TestIterateStorage() {
 	key2 := common.BigToHash(big.NewInt(3))
 	value2 := common.BigToHash(big.NewInt(4))
 
-	keeper := NewMockKeeper()
-	db := statedb.New(sdk.Context{}, keeper, emptyTxConfig)
+	raw, ctx, keeper := setupTestEnv(suite.T())
+	db := statedb.New(ctx, keeper, emptyTxConfig)
 	db.SetState(address, key1, value1)
 	db.SetState(address, key2, value2)
 
@@ -553,7 +595,11 @@ func (suite *StateDBTestSuite) TestIterateStorage() {
 
 	storage := CollectContractStorage(db)
 	suite.Require().Equal(2, len(storage))
-	suite.Require().Equal(keeper.accounts[address].states, storage)
+
+	ctx, keeper = newTestKeeper(suite.T(), raw)
+	for k, v := range storage {
+		suite.Require().Equal(v, keeper.GetState(ctx, address, k))
+	}
 
 	// break early iteration
 	storage = make(statedb.Storage)
@@ -565,6 +611,150 @@ func (suite *StateDBTestSuite) TestIterateStorage() {
 	suite.Require().Equal(1, len(storage))
 }
 
+func (suite *StateDBTestSuite) TestNativeAction() {
+	_, ctx, keeper := setupTestEnv(suite.T())
+	storeKey := testStoreKeys["testnative"]
+	transientKey := testTransientKeys[evmtypes.TransientKey]
+	memKey := testMemKeys[capabilitytypes.MemStoreKey]
+
+	eventConverter := func(event sdk.Event) (*ethtypes.Log, error) {
+		converters := map[string]statedb.EventConverter{
+			"success1": func(event sdk.Event) (*ethtypes.Log, error) {
+				return &ethtypes.Log{Data: []byte("success1")}, nil
+			},
+			"success2": func(event sdk.Event) (*ethtypes.Log, error) {
+				return &ethtypes.Log{Data: []byte("success2")}, nil
+			},
+		}
+		converter, ok := converters[event.Type]
+		if !ok {
+			return nil, nil
+		}
+		return converter(event)
+	}
+
+	stateDB := statedb.New(ctx, keeper, emptyTxConfig)
+	contract := common.BigToAddress(big.NewInt(101))
+
+	stateDB.ExecuteNativeAction(contract, eventConverter, func(ctx sdk.Context) error {
+		store := ctx.KVStore(storeKey)
+		store.Set([]byte("success1"), []byte("value"))
+		ctx.EventManager().EmitEvent(sdk.NewEvent("success1"))
+
+		transient := ctx.KVStore(transientKey)
+		transient.Set([]byte("transient"), []byte("value"))
+
+		mem := ctx.KVStore(memKey)
+		mem.Set([]byte("mem"), []byte("value"))
+
+		return nil
+	})
+	stateDB.ExecuteNativeAction(contract, eventConverter, func(ctx sdk.Context) error {
+		store := ctx.KVStore(storeKey)
+		store.Set([]byte("failure1"), []byte("value"))
+		ctx.EventManager().EmitEvent(sdk.NewEvent("failure1"))
+
+		transient := ctx.KVStore(transientKey)
+		suite.Require().Equal([]byte("value"), transient.Get([]byte("transient")))
+
+		mem := ctx.KVStore(memKey)
+		suite.Require().Equal([]byte("value"), mem.Get([]byte("mem")))
+		return errors.New("failure")
+	})
+
+	// check events
+	suite.Require().Equal(sdk.Events{{Type: "success1"}}, stateDB.NativeEvents())
+	suite.Require().Equal([]*ethtypes.Log{{
+		Address: contract,
+		Data:    []byte("success1"),
+	}}, stateDB.Logs())
+
+	// test query
+	stateDB.ExecuteNativeAction(contract, nil, func(ctx sdk.Context) error {
+		store := ctx.KVStore(storeKey)
+		suite.Require().Equal([]byte("value"), store.Get([]byte("success1")))
+		suite.Require().Nil(store.Get([]byte("failure1")))
+		return nil
+	})
+
+	rev1 := stateDB.Snapshot()
+	stateDB.ExecuteNativeAction(contract, eventConverter, func(ctx sdk.Context) error {
+		store := ctx.KVStore(storeKey)
+		store.Set([]byte("success2"), []byte("value"))
+		ctx.EventManager().EmitEvent(sdk.NewEvent("success2"))
+		return nil
+	})
+	stateDB.ExecuteNativeAction(contract, eventConverter, func(ctx sdk.Context) error {
+		store := ctx.KVStore(storeKey)
+		store.Set([]byte("failure2"), []byte("value"))
+		ctx.EventManager().EmitEvent(sdk.NewEvent("failure2"))
+		return errors.New("failure")
+	})
+
+	// check events
+	suite.Require().Equal(sdk.Events{{Type: "success1"}, {Type: "success2"}}, stateDB.NativeEvents())
+
+	suite.Require().Equal([]*ethtypes.Log{{
+		Address: contract,
+		Data:    []byte("success1"),
+	}, {
+		Index:   1,
+		Address: contract,
+		Data:    []byte("success2"),
+	}}, stateDB.Logs())
+
+	// test query
+	stateDB.ExecuteNativeAction(contract, nil, func(ctx sdk.Context) error {
+		store := ctx.KVStore(storeKey)
+		suite.Require().Equal([]byte("value"), store.Get([]byte("success1")))
+		suite.Require().Equal([]byte("value"), store.Get([]byte("success2")))
+		suite.Require().Nil(store.Get([]byte("failure2")))
+		return nil
+	})
+
+	stateDB.RevertToSnapshot(rev1)
+
+	// check events
+	suite.Require().Equal(sdk.Events{{Type: "success1"}}, stateDB.NativeEvents())
+	suite.Require().Equal([]*ethtypes.Log{{
+		Address: contract,
+		Data:    []byte("success1"),
+	}}, stateDB.Logs())
+
+	_ = stateDB.Snapshot()
+	stateDB.ExecuteNativeAction(contract, eventConverter, func(ctx sdk.Context) error {
+		store := ctx.KVStore(storeKey)
+		store.Set([]byte("success3"), []byte("value"))
+		ctx.EventManager().EmitEvent(sdk.NewEvent("success3"))
+		return nil
+	})
+
+	// check events
+	suite.Require().Equal(sdk.Events{{Type: "success1"}, {Type: "success3"}}, stateDB.NativeEvents())
+
+	// test query
+	stateDB.ExecuteNativeAction(contract, eventConverter, func(ctx sdk.Context) error {
+		store := ctx.KVStore(storeKey)
+		suite.Require().Equal([]byte("value"), store.Get([]byte("success1")))
+		suite.Require().Nil(store.Get([]byte("success2")))
+		suite.Require().Equal([]byte("value"), store.Get([]byte("success3")))
+		return nil
+	})
+
+	suite.Require().NoError(stateDB.Commit())
+
+	// query committed state
+	store := ctx.KVStore(storeKey)
+	suite.Require().Equal([]byte("value"), store.Get([]byte("success1")))
+	suite.Require().Nil(store.Get([]byte("success2")))
+	suite.Require().Equal([]byte("value"), store.Get([]byte("success3")))
+	suite.Require().Nil(store.Get([]byte("failure1")))
+	suite.Require().Nil(store.Get([]byte("failure2")))
+
+	// check events
+	suite.Require().Equal(sdk.Events{{Type: "success1"}, {Type: "success3"}}, ctx.EventManager().Events())
+}
+
 func CollectContractStorage(db vm.StateDB) statedb.Storage {
 	storage := make(statedb.Storage)
 	db.ForEachStorage(address, func(k, v common.Hash) bool {
@@ -572,6 +762,93 @@ func CollectContractStorage(db vm.StateDB) statedb.Storage {
 		return true
 	})
 	return storage
+}
+
+var (
+	testStoreKeys     = sdk.NewKVStoreKeys(authtypes.StoreKey, banktypes.StoreKey, evmtypes.StoreKey, "testnative")
+	testTransientKeys = sdk.NewTransientStoreKeys(evmtypes.TransientKey)
+	testMemKeys       = sdk.NewMemoryStoreKeys(capabilitytypes.MemStoreKey)
+)
+
+func cloneRawState(t *testing.T, cms sdk.MultiStore) map[string]map[string][]byte {
+	result := make(map[string]map[string][]byte)
+
+	for name, key := range testStoreKeys {
+		store := cms.GetKVStore(key)
+		itr := store.Iterator(nil, nil)
+		defer itr.Close()
+
+		state := make(map[string][]byte)
+		for ; itr.Valid(); itr.Next() {
+			state[string(itr.Key())] = itr.Value()
+		}
+
+		result[name] = state
+	}
+
+	return result
+}
+
+func newTestKeeper(t *testing.T, cms sdk.MultiStore) (sdk.Context, *evmkeeper.Keeper) {
+	encodingConfig := evmenc.MakeConfig(app.ModuleBasics)
+	appCodec := encodingConfig.Marshaler
+	cdc := encodingConfig.Amino
+
+	paramsKeeper := app.InitParamsKeeper(appCodec, cdc, sdk.NewKVStoreKey(paramstypes.StoreKey), sdk.NewTransientStoreKey(paramstypes.StoreKey))
+	subSpace, find := paramsKeeper.GetSubspace(authtypes.ModuleName)
+	require.True(t, find)
+	subSpaceBank, find := paramsKeeper.GetSubspace(banktypes.ModuleName)
+	require.True(t, find)
+	subSpaceEVM, find := paramsKeeper.GetSubspace(evmtypes.ModuleName)
+	require.True(t, find)
+	accountKeeper := authkeeper.NewAccountKeeper(
+		appCodec, sdk.NewKVStoreKey(authtypes.StoreKey), subSpace, ethermint.ProtoAccount, nil,
+	)
+	bankKeeper := bankkeeper.NewBaseKeeper(
+		appCodec, sdk.NewKVStoreKey(banktypes.StoreKey), accountKeeper, subSpaceBank, nil,
+	)
+
+	allKeys := make(map[string]storetypes.StoreKey, len(testStoreKeys)+len(testTransientKeys)+len(testMemKeys))
+	for k, v := range testStoreKeys {
+		allKeys[k] = v
+	}
+	for k, v := range testTransientKeys {
+		allKeys[k] = v
+	}
+	for k, v := range testMemKeys {
+		allKeys[k] = v
+	}
+
+	evmKeeper := evmkeeper.NewKeeper(
+		appCodec, testStoreKeys[evmtypes.StoreKey], testTransientKeys[evmtypes.TransientKey], subSpaceEVM,
+		accountKeeper, bankKeeper, nil, nil,
+		"", nil,
+		allKeys,
+	)
+
+	ctx := sdk.NewContext(cms, tmproto.Header{}, false, log.NewNopLogger())
+	return ctx, evmKeeper
+}
+
+func setupTestEnv(t *testing.T) (sdk.MultiStore, sdk.Context, *evmkeeper.Keeper) {
+	db := dbm.NewMemDB()
+	cms := rootmulti.NewStore(db, log.NewNopLogger())
+	for _, key := range testStoreKeys {
+		cms.MountStoreWithDB(key, storetypes.StoreTypeIAVL, nil)
+	}
+	for _, key := range testTransientKeys {
+		cms.MountStoreWithDB(key, storetypes.StoreTypeTransient, nil)
+	}
+	for _, key := range testMemKeys {
+		cms.MountStoreWithDB(key, storetypes.StoreTypeMemory, nil)
+	}
+	require.NoError(t, cms.LoadLatestVersion())
+
+	ctx, keeper := newTestKeeper(t, cms)
+	keeper.SetParams(ctx, evmtypes.Params{
+		EvmDenom: "uphoton",
+	})
+	return cms, ctx, keeper
 }
 
 func TestStateDBTestSuite(t *testing.T) {