+++ CHANGELOG.md
@@ -36,6 +36,12 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
+## Unreleased
+
+### Features
+
+* (evm) [#328](https://github.com/crypto-org-chain/ethermint/pull/328) Support precompile interface.
+
 ## [v0.19.3] - 2022-10-14
 
 * (deps) [1381](https://github.com/evmos/ethermint/pull/1381) Bump sdk to `v0.45.9`